refactor: extract generic prompt helpers
